Shift products models from main to product app
